Update patch set 5
Patch Set 5: I would prefer that you didn't merge this

The code looks fine but please put a summary of the change in the commit message; the blueprint search fails in this case for some reason.  This is a Gerrit feature and does not carry in to the git commit message itself.

Patch-set: 5
Label: Code-Review=-1
Label: Workflow=0