Update patch set 1
Patch Set 1:

Agreed.

I did that in configure because neutron apparently does everything there - and the code for creating the db was there.

I will move the code in init_neutron.

I hope it is ok to have a check for creating the db and performing the migration only if q-svc is enabled, as the database should not be created on hosts where only agents are deployed.

Patch-set: 1
Label: Code-Review=0