run_tests.sh can not install Python package
On Ubuntu 12.04, while running run_tests.sh in /opt/stack/nova,
it will install the virtual environment firstly. The Python packages
listed in requirements.txt and test-requirements.txt should be
installed properly. But we need to install 3 packages to pass it:

libxslt1-dev
libmysqlclient-dev
libpq-dev

It is better to install them while installing nova in devstack.

Change-Id: I7ee9427513da98498f39cc16649c4aa7f22824a4