Update patch set 3
Patch Set 3:

it's definitely weird to me to land something in devstack that's not at minimum in stackforge. Having this dependent on an fully external git tree which we can't gate on is really odd.

I read the nova reviews for this though that wanted to wait on devstack... so I'm a bit torn.

Patch-set: 3
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0