Update patch set 3
Patch Set 3: (4 inline comments)

I started to answer the stackforge question inline. Basically I wanted to be on Stackforge first. Being new to OpenStack in general, I asked one guy on IRC who told me that the project was not big enough to justify a Stackforge project. It makes sense regarding the existing ones. In anyway, I am open to this and I guess Monty is the maintainer of it, so I'll ask him.

Thanks for the feedback.

Patch-set: 3
Label: Code-Review=0