Update patch set 4
Patch Set 4:

Bob, I've had a bit more time to digest this, I understand it's basically just to get os_PACKAGE defined; as it stands there are a few loose ends I'd like to clean up.

Based on the string in GetOSVersion you'll wind up with:

 os_VENDOR=XenServer
 os_CODENAME=xenenterprise
 os_UPDATE=0-70446c
 os_RELEASE=6.2
 os_PACKAGE=rpm

Then GetDistro will produce:

 DISTRO=XenServer-6.2.0-70446c

I think this needs to be cleaned up a bit to be more useful.  Simply adding 

 "$os_VENDOR" =~ (XenServer)

to the existing Red Hat/CentOS check will produce:

 DISTRO=rhel6

which may or may not be a good thing.  Alternatively, an additional elif block specific to XenServer would let you make it:

 DISTRO=xen6
 DISTRO=xen6.2

or whatever.  I'd suggest 'xen6' if the minor release is used the same way it is in RHEL.


Also, please put a bit more of the reason for this in the commit message.  It's helpful when looking at things offline and LaunchPad bugs are not available.

Patch-set: 4
Label: Code-Review=0
Label: Workflow=0