Update patch set 2
Patch Set 2:

Sorry - I realized that we needed to use xenbr0 rather than br100 in the common case.  I was going to add a patch to XAPI to allow it to use an already created bridge; but that wouldn't have worked because eth0 is already on xenbr0 so can't go on br100.

As such the networking has to differ from libvirt a little bit, and the need for an update to this patch.

Patch-set: 2
Label: Code-Review=0