Update patch set 5
Patch Set 5:

I'm not sure I understand why this belongs in DevStack. It's a) for old versions of a hypervisor and b) very specific to nova-compute. Shouldn't it be included with the driver that needs it?

Patch-set: 5
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0