blueprint ceph-storage-backend
This will allow the automated creation of a single node ceph mon/osd and configure glance/cinder to use it for storage. Tested with ceph 0.56 but in theory is should work on any Bobtail/Cuttlefish release of ceph. To enable add ceph to the list of ENABLED_SERVICES in your localrc.

Change-Id: Ia4cfbc4b730e77485ddedbc48c829b7a474883c1