Colorize Neutron log output and refactor log setup code
Bug 1214616

This patch adds colors to on-screen Neutron log output in the same way nova,
cinder, and heat engine do.
To this aim, logging configuration instructions are moved into ./functions.
The reason for this refactoring is that these instruction are the same for each
project, with the only exception of the target configuration file.

Change-Id: Idf0d1b842bb9ab046c9ef826de1dfc55b3f1df9d