Update patch set 1
Patch Set 1: Do not merge

I don't know what mageia is (oh, a Mandriva fork). Please convince me that adding it is worth the effort as it will not be supported by any of the core devs and will potentially be broken by the next patch that comes along.  This is the first I've heard of the need for it so some demonstration of a community of support would be good.

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2
Label: Workflow=0