Update patch set 1
Patch Set 1:

This is part of an effort to make Mageia a major distribution for computational astrophysics.  The first step is to have openstack and devstack run cleanly on the next release of Mageia, and then combine this with some of the astrophysical tools that have been added.  To do this, we've made a number of bug fixes and enhancements to supermin and libvirt, and I've submitted a talk to the Hong Kong conference about everything we've had to do to get openstack to work.

It's possible for us to keep the patches on a separate tree, but I think that adds more work for everyone.

Patch-set: 1
Label: Code-Review=0