Update patch set 1
Patch Set 1:

Dear Dean,

I'm pretty surprised of your reaction. You don't know Mageia but please consider the following point.

- Mageia is not Mandriva
We did fork Mandriva because of some bad strategic decisions taken by them. Since 3 Years we are making our own at 100%.

- Mageia is a 100% community distribution
As we can look here (https://wiki.mageia.org/en/Org), we are running as a community. After 3 successful releases, I can say that it this model works well.

- Mageia is more popular than you thought

That's not an absolute meaning but on the last 12 month we are number 2 on DistroWatch.

Linux.com listed us as the one of the 5 most important project of 2012 : http://www.linux.com/learn/docs/682469-the-5-most-important-linux-projects-of-2012

- Mageia is not only a Distro for Desktops
Mageia is not designed for being a distro for desktop even if it's nice to run this way.  Really, we are working hard to have a good product that people like. 

- Magiea is friendly with upstream
Openstack is a strong demand this days and we want to satisfy it. We aim at being friendly with upstream projects and that's why Joseph came to you to add Mageia in Openstack. We don't aim at keeping the work for us and apply patches just for the fun. If we had to add some in our packages, upstream team would be informed to discuss it and integrate it if it worth.

Please approve this addition, that would help us to grow and continue our way to get a nice community distribution.

Regards,
Erwan

Patch-set: 1
Label: Code-Review=0