Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

I think this should be kept to the minimum size required by the tests to complete successfully.

If the concerns about the total number of IPs available are due to the parallelization of the tests run, then I agree with Sean that this should be a change in devstack-gate and not in devstack.

Patch-set: 1