Update patch set 4
Patch Set 4:

This is going to slow devstack-gate runs down because the package won't be cached like all the other debs.

The nodepool setup scripts (previously in devstack-gate, soon to be removed) are in the openstack-infra/config repository in ./modules/openstack_project/files/nodepool/scripts/

They are used to cache things that devstack uses, such as debs and images for our gate runs.  It's actually possible that merging this might even break those setup scripts if the package name is different than what's in the normal ubuntu archive (I haven't checked).  At any rate, it would be good to add the archive in those image setup scripts so that it can cache the correct deb.

Finally, I wonder if this should check to see if the archive is already in place and short-circuit the keyserver and apt_update if it is.

Patch-set: 4
Label: Code-Review=0
Label: Workflow=0