Set scenario.large_ops_number in tempest.conf
The large_ops test tries to catch any performance issues when running
large numbers of operations at once, in this case launching instances.

Set to 25, this is the number of instances that the large_ops test will
try to spin up at once, this defaults to 0 in tempest itself because
there is no good universal default, but for devstack there is.

25 was picked somewhat randomly as number that works pretty well.

Part of regression test for bug 1199433

Change-Id: I459717b849f9b6d180a7956c0ff470cefe7c1ff6