Update patch set 3
Patch Set 3:

yeah, ideally we would use the cirros image here, rather than forcing one to manually add it.  I briefly looked at using qemu-img to be able to convert the cirros img file to a vmdk file that could be used by the VMwareVCDriver, but this did not work, at least with the glance options I used.  If we could get this working it would be great, but I don't see it as a blocker.

Patch-set: 3
Label: Code-Review=0
Label: Workflow=0