Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

(3 inline comments)

I'd like for this to better match the existing practices in lib/keystone.  We already have KEYSTONE_IDENTITY_BACKEND that duplicates the KEYSTONE_AUTHENTICATION functionality, although it is not properly defaulted or vetted.

All of the existing 'backend' variables use the second component of the module path name as the middle part of the variable name.  Please do that here too.

Is this to support functionality that will be included in the Havana release?

Patch-set: 4
Label: Code-Review=-1
Label: Workflow=0