Update patch set 15
Patch Set 15: I would prefer that you didn't merge this

(16 inline comments)

This is only a partial review as I was not able to get the tr-* services to start in an otherwise fresh default DevStack install.  This is primarily due to attempting to create the keystone artifacts before keystone is started.

Patch-set: 15
Label: Code-Review=-1
Label: Workflow=0