Update patch set 1
Patch Set 1:

I think it is ok using the same signing_dir between the 3 api servers as all the other info in the keystone_auth section is the same. But it would be good for Steve and Steven to ack that.

Patch-set: 1
Reviewer: Gerrit User 4715 <4715@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0