Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

there is no feature freeze per say on devstack, we cut stable some time late in the cycle. I think this is fine to have, but I don't want to merge it until after the tempest changes are merged, because I think we expose users to a race in the gate otherwise, which I'd like to avoid. So +1 but don't approve until the tempest change is in.

Patch-set: 3