Clean up automated changes to requirements
Some of us like to reuse sandboxes, and keep them up to date.  This
is very difficult to do if devstack leaves modifications to
requirements.txt files after a run, since 'git pull' may refuse to
overwrite those changes.

This modification has devstack reset the sandbox after installation
when it made a requirements.txt update, to leave it in a clean state
again.

Change-Id: Ia2d928ade8141b59b56a2c4548d760bf6911a3e5