ML2 can't seem to load type drivers
fixes bug: 1221781

When ML2 plugin is initialized using Devstack, jsonrpclib is not
installed. This prevents some type drivers from starting up.

A fix is introduced such that when ML2 plugin is configured using
devstack, jsonroplib is installed. This package needs to be installed
only on the server node. Hence the package is installed in routine
neutron_plugin_configure_service().
Initially, it is used by Arista mechanism driver.
In future, it may be used by other drivers as well.

Change-Id: Iafceeaed872192c2c4d2101d05ad68183b60ee49