Update patch set 4
Patch Set 4:

I'm working in a virtualbox managed by vagrant. The file ownership is mapped to my local user outside of the VM, and the chown/chmod operations to change stuff to be owned by the vagrant user is failing.

I'm not sure about all of the things install does, but I don't think that's going to fit with how chown/chmod are being used here. It's for things like changing the ownership of egg-info directories after running "sudo python setup.py develop", since on a local filesystem that leaves the egg-info directory and its contents owned by root, which may mean an sdist command fails later and breaks tox.

Patch-set: 4
Label: Code-Review=0