Update patch set 1
Patch Set 1:

I am still trying to check if this fix actually works, but because I don't know the actual root cause of this bug yet, the best I can do is recheck this patch a bunch of times.  So far I have not been able to reproduce the bug in 7 attempts and I will keep trying until I hit 20 attempts.  After which point I think it may be safe to say that this fix is working.

Patch-set: 1