Add root CONF_DIR variable.
Add a root CONF_DIR variable so that config files for the various
OpenStack services can be grouped together under one dir that isn't
shared with the rest of the system like /etc. This is useful for
capturing the config files in devstack-gate after testing.

Note that horizon, apache, and tempest are not changed as their config
file dirs are not generic like the others.

Change-Id: I3462e3089eb121cd8590fca50f9d1f5cacc8d2a1