Update patch set 1
Patch Set 1:

Thanks for the reviews!  This raises an interesting question for us on how we should approach this issue.  Our model to plug in VMs to our virtual network is a bit unique and does not exactly fit the model of current Nova vif drivers, making it difficult to submit upstream (though we  certainly plan to work on this).  I'm wondering if it's possible to add a hook in neutron_thirdparty to configure nova configuration file for something like this.  Right after neutron_plugin_create_nova_conf, we can call neutron_plugin_thirdparty_create_nova_conf.  We would be more than happy to submit a patch for this if the core devstack maintainers agree that it's a reasonable solution for this issue.  Looking forward to hearing back!

Patch-set: 1
Reviewer: Gerrit User 156 <156@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0