Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

Jamie, the blueprint is inadequate, basically it says using SSL is good, no argument there. But what's missing is:

1) Why isn't SSL possible at the moment?

2) What needs to be added to support SSL?

3) How do you plan to implement the missing functionality?

4) In what context is this valid? (i.e. only devstack).

5) How would the issue be solved outside the target context (i.e. outside devstack).

6) What are the underlying assumptions/requirements? (i.e. this only works with OpenSSL and it's configuration methodology).

7) Is this a general purpose solution or a one-off stepping stone to a better design?

8) What are the inherent problems with this approach? The design doesn't have to solve the problems but it should at least enumerate the issues so they can be taken into consideration and provide a historical context.

I can derive a fair amount of this from the patch set, but that's backwards, rather given the design I want to know if the patch set implements the design.

Patch-set: 4