Update patch set 4
Patch Set 4: No score

I'm confused. A quick perusal of the patch set suggests what is occurring is the distribution of the CA cert. But perhaps I'm wrong, that's why I wanted to blueprint to clearly the goal and solution.

My understanding of certmonger is it's functionality is the generation and renewal of certs. certmonger has to the best of my knowledge nothing to do with the distribution of certs, that's something else entirely.

Jamie please fill out the blueprint so we can all operate from the same set of assumptions without confusion.

Patch-set: 4
Label: Code-Review=0