Nicira plugin: do not die if NVP gateway IP is missing
Devstack should not die if the parameters for establishing a
management connection to the NVP gateway appliance are not
provided, as this is not critical for devstack itself, and
also the user might have already configured connectivity
outside of devstack.

This patch also avoid using $PUBLIC_BRIDGE to this aim.
With the nicira plugin, this bridge is created only if
Q_USE_DEBUG_COMMAND is enabled; this is not a requirement for
running devstack unless one wants to run exercises.

Bug #1227750

Change-Id: Ib58738a578c46f2183d503cabfdc6039bfbeb702