Neutron nicira plugin: set up connectivity with public network
When spinning devstack with the Neutron and the Nicira plugin, the public
bridge is not configured since the external network will be implemented
on the NVP gateway appliance.

This patch therefore configures an IP address for the devstack interface
on the NVP gateway's uplink network thus providing IP connectivity with
the Neutron external network configured by devstack.
The configured IP address is the external network's gateway, in
accordance with similar setting for OVS-based plugins.

This patch also ensures the IP address is configured when devstack
is shut down.

Bug #1227757

Change-Id: I971eb75548af23c2c77506d48eabd49f2ddb7348