Update patch set 4
Patch Set 4:

Nalin Dahyabhai is the auther of certmonger, I sought clarification from Nalin on IRC as to the capabilities of certmonger.
I'm preserving the IRC conversation here for reference.

(12:11:45 PM) jdennis: nalin: question about certmonger functionality, my understanding was certmonger implemented the generation and renewal of certs, but it has no functionality
(12:12:05 PM) jdennis: for the distribution and installation of said certs, correct?
(12:13:33 PM) jdennis: in other words if certmonger is managing a CA cert some other mechanism has to be available to get that cert and it's trust into an OpenSSL ca bundle or a NSS database, yes/no?
(12:26:08 PM) nalin: jdennis: it stores certificates directly in NSS databases, so you shouldn't need an intermediate location for that. but for files, the assumption is that the certificate is the only content in the file, so you'd need to manually manage it if it was part of a bundle file
(12:27:44 PM) jdennis: nalin: tnx
(12:28:15 PM) nalin: sure
(12:28:40 PM) jdennis: nalin: wait a minute I think there might have been some confusion with my question
(12:30:00 PM) nalin: if by distribution you mean to other systems once they're stored to local files, then no, it doesn't know how to do that
(12:30:14 PM) nalin: (local files or NSS databases, that is)
(12:30:20 PM) jdennis: nalin: lets say certmonger generates a CA cert, it get's stored someplace (file, NSS db, etc.), but let's say you've got 10 other systems which need the CA cert in one or more locations, certmonger can't help with that right?
(12:31:06 PM) nalin: not directly, no. the most it can currently do is fire off a designated command after it stores the new certificate to disk
(12:31:47 PM) jdennis: that's primarily to handle renewal, correct?
(12:32:21 PM) nalin: the option was added to be used for that case, but it's not restricted to that case
(12:32:53 PM) jdennis: right, I could see it being used to initially populate
(12:34:13 PM) nalin: ... the list of trusted certs in VMs/containers, from the host?
(12:35:50 PM) jdennis: that's one use, another is cooperating services on different nodes using X509 auth with a common CA
(12:36:45 PM) nalin: are those nodes driving the renewal of their own certs, and if so, where would they be distributed to?
(12:37:47 PM) jdennis: that's one case, but the case I was specifically thinking of was renewal of the CA cert and it's distribution
(12:39:45 PM) nalin: that's going to be tricky, since we normally expect that deleting the old certificate is harmless if we're attempting to store the new one, and i don't think you'd generally want to do that with a CA cert
(12:41:27 PM) jdennis: but that's the province of the triggered external command run on renewal to get correct, not a consideration for certmonger
(12:42:07 PM) nalin: if you're pushing it from the location certmonger saves it to, out to somewhere else, sure. i'm talking about that first copy
(12:42:45 PM) jdennis: ah ok, understood

Patch-set: 4