An example of Certificate Generation with devstack
This uses the in-built CA generator rather than certmonger because
1. As far as i can see certmonger will ONLY issue self-signed unless
   attached to an external CA.
2. I would have then had to figure out how to make sure the certmonger
   process was up and running before requesting.

However the basic process is the same. In the event that you don't get
the certificates you expect for an enabled cert then you should generate
some new ones. Whether we used the inbuilt OpenSSL or Certmonger is then
just figuring out how to do that abstraction in bash.

Because we just set the environment variables that the rest of devstack
expects the whole thing will work in much the same way as the user
inputted variables.

Change-Id: I3a80481743541822581e694cd6c37d4b49e43b6f