Nicira plugin: do not die if NVP gateway IP is missing
Devstack should not die if the IP and prefix len for establishing
a connection to the public network are not provided.
In this case, the public gateway IP address used to configure
Neutron's public network should be used, together with the prefix
length of the public network's CIDR.

This patch also ensures $PUBLIC_BRIDGE is created, even if
Q_USE_DEBUG_COMMAND is disabled. Finally this patch also adds
the teardown operation for restoring the original IP addresses
on the interface used for connectivity to the public network
implemented on the NVP gateway.

Bug #1227750

Change-Id: Ib58738a578c46f2183d503cabfdc6039bfbeb702