xenapi: enable user interaction with stack.sh
In XenServer, devstack runs inside a virtual machine. This makes it hard
for the user to interact with stack.sh. This change will create upstart
jobs, that run stack.sh on hvc0 - so user can interact with the script.
As devstack finished, the usual login prompt will be displayed.

As devstack's output is also logged, tee is used. This caused
difficulties, because it was buffering the last line, which used to
detect if the operation succeeded. Now a separate stamp file is created,
so tee's buffering should not be an issue

Change-Id: I7eb12bd74746cc7a1db3aa9fd68ece645a50001d