xenapi: enable user interaction with stack.sh
In XenServer, devstack runs inside a virtual machine. This makes it hard
for the user to interact with stack.sh. This change will create an upstart
job, that runs stack.sh, and enables console, so user can interact with
it.

Logging the output is also disabled, stamp files are used instead to
detect the script run status.

standard devstack logging should be used instead.

Change-Id: I7eb12bd74746cc7a1db3aa9fd68ece645a50001d