Update patch set 1
Patch Set 1:

Sean, thank you for you review. I was not too familiar with Gerrit commands, sorry for "no bug".

Currently in Ironic team we are trying to make some tempest tests so if we do not expose a http://host:port endpoint now we have to update tempest tests later.

Also there are some API discovery features that cannot be accessed (and tested) with the http://host:port/v1 endpoint.

Patch-set: 1