Make unstack.sh work when run as root
Make unstack.sh check whether it's running as root, and if so, use the
same approach as stack.sh in order to switch to the $STACK_USER account,
and re-exec itself.  In contrast to stack.sh, we can skip the sudo and
$STACK_USER account setup code, because this should have already been
taken care of via a stack.sh run (and if it hasn't, presumably there is
nothing worth "unstacking").

Change-Id: If42ad6270d60f4b01b1587032448654699d4c812