force run_process to really spawn a bash process
Before this change ceilometer/sar process are killed by stack.sh in gate

gate use 'USE_SCREEN=False' for devstack, this run service process in
some bash subshell instead of a screen.

The ceilometer service or the sar command are started directly without embeded shell
command: 'screen_it ceilometer-api "ceilometer-api -d -v ..."',
and this have as effect to not add a additionnal bash (subshell) process between
the stack.sh process and the ceilometer/sar process.

And when stack.sh kill all directly attached process (clean function),
it kill the with ceilometer and sar too.

This is because the 'bash -c "only_one_command"' use exec to start the command
and does not execute the command as a bash script.

This patch use "{ $command ; }" to be sure that bash don't exec the command
passed to '-c' but execute it as a shell script.

Fixes bug #1234254
Fixes bug #1238482

Change-Id: I965506d9105608c936bcbcb5913d1e154bedaad3