Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

I think Medhi's fix is the way to go. It took us days to find the original issue in Ceilometer, as we don't need to do any sort of "cd /foobar &&" to run it. Any service that will have the same non-requirement will encounter this problem.

This run_process function has a bug, we just encountered it. It is better to fix it than to have each of the service papering over it.

Patch-set: 3
Label: Code-Review=+1