Update patch set 3
Patch Set 3:

Why do we need to workaround the bug of run_process in each screen_it call ? the role of run_process is clear, create a simulation of double fork to detach a process.

If this is the "true;$command" that is not sexy, we can write '{ $command ; }', it works too and it look less hackish.

Patch-set: 3
Label: Code-Review=0