Update patch set 1
Patch Set 1:

There's nothing on the website or in the README.md which says stack.sh should or shouldn't be run as root.  stack.sh's current behaviour is to fairly gracefully handle running as root, so it seems entirely reasonable for newcomers to assume it should work.  If it really is officially unsupported, then it's a very misleading state of affairs for newcomers :-/  In that case it should at very least barf with a "please don't run me as root" error message.

I totally share your concerns about HTTP_PROXY etc. so I'm not going to argue against the idea of removing root support in either the short or long term.  However I think we need to clarify the status quo and set users' expectations correctly ASAP.

So IMHO there are only three reasonable ways forward:

1. *immediately* remove support for running as root, and update the website to explain how to correctly set up sudoers, or

2. accept short-term fixes for current breakage when running as root, but maintain a medium or long term target of removing support for running as root, or

3. decide to maintain support for running as root indefinitely.

What I'd hate to see happen is:

4. my fixes for the breakage get rejected, but partially broken root support is left in for any significant period of time.

On a similar topic, it's not very clear from the website (to me at least) which distro platforms are supported or where support can be found, but that's another can of worms which I'll open elsewhere ...

Patch-set: 1