Update patch set 4
Patch Set 4:

Sean, with all my due respect, the fact that you do not understand unfortunately don't imply that you are right. You are basically adding -2 without even arguing, which is not a good thing to do.

Therefore, I'll try to explain.

There's two cases when using "bash -c":
#1 When using "bash -c 'command'", the bash process belongs to the same process group that the process that called "bash -c".

#2 When using "bash -c 'command1; command2'", the bash process starts a new process group to run these commands. This also happens when using "bash -c '{ command; }'" as this starts a new process group.


In devstack, there's a clean up procedure that kills all jobs started by the script using "jobs -p". The -p option lists all the processes that are part of the same group process of the shell running "jobs -p", in this case the script.


Some projects in devstack runs the commands "cd foobar && foobar", which is translated into case #2. In such case, "jobs -p" does not return the PID of these, and they are not cleaned out by the killing spree.

However, Ceilometer has no need to change the directory before running, so we simple asked "ceilometer-api" to be ran. However, this translates to case #1. And in the end, that makes Ceilometer part of the same process group as the script, therefore listed in the "jobs -p" output and it is killed as the environment is cleaned. Rendering Ceilometer non functionnal.

Nothing in the prototype or documentation of this function indicates that the command should be composed of at least two components, that's why Mehdi's proposing to fix this function to work in _every_ cases and to to make all its call fall into the case #2. Fixing our the Ceilometer problem in a clean way, and avoiding this pitfalls for future users.

Patch-set: 4