Update patch set 7
Patch Set 7: I would prefer that you didn't merge this

great improvment, especially to configure vxlan section of ML2 plugin, when it is used with linuxbridge. 
I had an issues, with a multi-host config. We are used to only enable services q-agt,n-cpu on compute node, but we have to had "enable services neutron" so that ml2_conf.ini will be populated with the [neutron] section.
"enable services neutron" is not mandatory on the controller.

Patch-set: 7
Label: Code-Review=-1