Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

so, this is good... however I think that we should build a job to sanity check things for this automatically before this is merged, so we can enforce it on all new patches at some point. Can you work with #openstack-infra on that?

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0