Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

so I -2ed to get an explanation, which I don't think is out of line. The point of a core reviewer is to ensure they are comfortable with what's going into the tree. What it looks like the issue is that ceilometer isn't calling out of the git tree but out of /usr/local instead (unlike other services).

So how does this impact the ability to C-c the process, change code in the git tree, and start it again with a command and taking the updates in the code? If this is a new model for processes we should probably apply it across the board instead of piece meal on a single project.

Patch-set: 4
Label: Code-Review=-1