Update patch set 7
Patch Set 7:

Kyle, you actually can change any file on the system using this, the group determines _if/when_ a particular section is evaluated, not which files it can manipulate.  That said, it appears to not have quite the if/then capability that you are looking for here.

To handle the MechanismDrivers, as it sits you would just have to have all of them that you expect prepared and let the config files be modified even if the driver is not loaded.  I gather that the selection of MechanismDrivers is a start-time decision and the requirement that all of local.conf be consistent with the configured services seems reasonable to me.  Am I missing anything here?  It does seem like the ability to select a different local.conf file to use on the stack.sh command line might be useful now.

Matthieu, multi-host has always been a weakness in DevStack as the localrc has always needed to be custom for each host.  That is still the case here, and as I mentioned above, the group name doesn't have to match the owner of the file you are modifying.  If you need to modify ml2_config.ini and only q-agt is enabled, then put q-agt as the group name on those hosts.

Patch-set: 7