Update patch set 3
Patch Set 3:

@ayoung: I tried this out and it works, is your concern just about the way KEYSTONE_ADDITIONAL_EXTENSIONS are specified? 

There are (most of the time) two main steps to enabling an extension: 1) adding it to the pipeline, and 2) running db_sync. 

The map option provides a nice way to be specific about what to provide to the pipeline and the db_sync command.

Would your list suggestion look like: 
KEYSTONE_ADDITIONAL_EXTENSION=oauth1,endpoint_filter?

If that is the case, then we would be assuming a _extension suffix would work for all cases (to perform step 1).

Patch-set: 3