protected_hardlinks should be disabled
The newer libguestfs as it used by nova not compatible with the
hard link protection.

This option is disabled on most systems, but it is enabled on f19 by
default.

protected_hardlinks adds an extra unusual owner check for hard link creation,
may be it is too protective.

Until the latest nova/libvirt/libguestfs combination is not
protected_hardlinks friendly, it should be disabled.

See also: https://www.kernel.org/doc/Documentation/sysctl/fs.txt

Change-Id: I9ea90e0b76e4be9d4bd9d722fd65244a3c114949