Update patch set 1
Patch Set 1:

Any more thoughts on this?  Rather than the s/POSIX // in patch set 1, I could do s/POSIX /bash /.  That seems a bit overkill to me given that bash is explicitly mentioned two sentences later, but if that's the difference between +1 and -1 then I'm happy to do that.  The main thing is that it doesn't claim to be a POSIX shell script, because that's plain wrong.

Patch-set: 1