Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

I'm fine with this, it does point to the need for a general sweep through comments and text again before stable/havana gets cut.  Thanks for the time doing a bunch of this already Adam.

A style checker would be great but I can't find anything useful for shell to start with and a bunch of our issues are in comments anyway.  If I get a chance I'll work on something but it'll be in Perl... ;)

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0