Update patch set 2
Patch Set 2:

Now I see three situation:
 - You have an already installed prettytable which does not have this issue (probably it is a distro package).

 - If your already installed version is in the expected range prettytable>=0.6,<0.8 , but not 0.7.2 the
 the original code fails without a real reason.
 
 - If your already installed version is out of range (prettytable>=0.6,<0.8) probably you will see an issue.
This situation is not covered by this change, and it was an issue before this change as well.
A possible solution is enforcing the version range here, but it will not be in sync with global requirements in the future.

- pip by default installing the latest 0.7.2 package if you do not have an already installed one, or your already had 0.7.2. In this case both the old and new code working properly. 

The real question is what will happen when the 0.7.x (x>2) released.
If >=0.7.2 will have the same issue, it will cause gate failure, as with the original code.

Our options are:
 - Enforcing the installation of v0.7.2:
    - We will upgrade(or downgrade) package from the accepted range. It is not a necessary operation. We might not update to new version when it is possible.
 - Touch more files permission (every version) even if it is not required.
    (Probably it was not an issue with the older versions.)
 - We are risking a gate issue if the new version will have the same issue. (This was the situation before and after this change.)
This option expects, prettytable will be fixed in the newer version.

Would you like to change the original strategy ? 
Which one do you prefer  ?

Patch-set: 2
Label: Code-Review=0