Update patch set 1
Patch Set 1:

If we want to get really picky, HEAT_CREATE_TEST_IMAGE is never defaulted anywhere so it relies on the user setting it to exactly "True" in the local config.  It would be clearer to either relax the test to possible -n / -z or use trueorfalse() to normalize the common user setting values.

Also, it may be my preference for using '[[' that has '==' creeping in to the conditionals even when '[' is used.  Since we're specifically targeting bash >= 3.x  I'd rather use '[['.

Patch-set: 1
Label: Code-Review=0